requirements.txt
functions-framework
google-cloud-firestore
requests
beautifulsoup4
pytz
telethon
google-cloud-secret-manager
google-cloud-storage

